+++ src/services/hal/camera_solution_proxy.cpp
@@ -346,7 +346,7 @@ bool CameraSolutionProxy::luna_call_sync(const char *func, const std::string &pa
     int64_t startClk = g_get_monotonic_time();
     luna_client->callSync(uri.c_str(), payload.c_str(), &resp, COMMAND_TIMEOUT);
     int64_t endClk = g_get_monotonic_time();
-    PLOGI("response %s, runtime %lld", resp.c_str(), (endClk - startClk) / 1000);
+    PLOGI("response %s, runtime %lld", resp.c_str(), (long long int)((endClk - startClk) / 1000));
 
     json j = json::parse(resp);
     if (j.is_discarded())
